Move Content.Server.Eye to Shared
