--- Content.Server/Eye/Blinding/ActivatableUIRequiresVisionComponent.cs
@@ -1,8 +1,0 @@
-namespace Content.Server.Eye.Blinding
-{
-    [RegisterComponent]
-    public sealed partial class ActivatableUIRequiresVisionComponent : Component
-    {
-    }
-}
-
