+++ Content.Shared/Eye/Blinding/Systems/ActivatableUIRequiresVisionSystem.cs
@@ -1,18 +1,14 @@
-using Content.Shared.Eye.Blinding;
 using Content.Shared.UserInterface;
-using Content.Server.Popups;
 using Content.Shared.Eye.Blinding.Components;
-using Content.Shared.Eye.Blinding.Systems;
-using Robust.Shared.Player;
-using Robust.Server.GameObjects;
+using Content.Shared.Popups;
 using Robust.Shared.Collections;
 
-namespace Content.Server.Eye.Blinding;
+namespace Content.Shared.Eye.Blinding.Systems;
 
 public sealed class ActivatableUIRequiresVisionSystem : EntitySystem
 {
-    [Dependency] private readonly PopupSystem _popupSystem = default!;
-    [Dependency] private readonly UserInterfaceSystem _userInterfaceSystem = default!;
+    [Dependency] private readonly SharedPopupSystem _popupSystem = default!;
+    [Dependency] private readonly SharedUserInterfaceSystem _userInterfaceSystem = default!;
 
     public override void Initialize()
     {
